+++ src/Features/ERDDAP/hooks/index.ts
@@ -1,4 +1,4 @@
-export { usePlatforms, useForecast, useForecasts } from "./buoyBarn"
+export { usePlatforms, useForecast, useForecasts, useForecastMeta } from "./buoyBarn"
 export { useDataset, useDatasets } from "./tabledap"
 export { UsePlatform, UsePlatforms } from "./BuoyBarnComponents"
 export { UseDataset, UseDatasets } from "./TableDAPComponents"